@@ -9,35 +9,35 @@ import com.jeanloth.project.android.kotlin.domain_models.entities.Command
 object DataMock {
 
     val articles = listOf(
-        Article(name = "Pain au beurre 1kg", price = 20.0),
-        Article(name = "Pain au beurre 500g", price = 10.0),
-        Article(name = "Chocolat", price = 10.0),
-        Article(name = "Rolls Kanèl", price = 10.0),
-        Article(name = "Petits pain", price = 0.50),
-        Article(name = "Petits pain x 20.0", price = 10.0),
-        Article(name = "Pain au choco x 10", price = 15.0),
-        Article(name = "Flan au coco", price = 10.0),
-        Article(name = "Velouté", price = 12.0),
-        Article(name = "Pommes cannelles", price = 15.0),
+        Article(label = "Pain au beurre 1kg", price = 20.0),
+        Article(label = "Pain au beurre 500g", price = 10.0),
+        Article(label = "Chocolat", price = 10.0),
+        Article(label = "Rolls Kanèl", price = 10.0),
+        Article(label = "Petits pain", price = 0.50),
+        Article(label = "Petits pain x 20.0", price = 10.0),
+        Article(label = "Pain au choco x 10", price = 15.0),
+        Article(label = "Flan au coco", price = 10.0),
+        Article(label = "Velouté", price = 12.0),
+        Article(label = "Pommes cannelles", price = 15.0),
     )
 
     val articleWrappers = createWrapperList(articles)
 
     val articleWrapper1 = ArticleWrapper(
-        article = Article(name = "Pain au choco", price = 15.0)
+        article = Article(label = "Pain au choco", price = 15.0)
     )
     val articleWrapper2 = ArticleWrapper(
-        article = Article(name = "Rolls Kanèl", price = 10.0),
+        article = Article(label = "Rolls Kanèl", price = 10.0),
         statusCode = ArticleWrapperStatusType.DONE.code
     )
 
     val articleWrapper3 = ArticleWrapper(
-        article = Article(name = "Chocolat", price = 12.0),
+        article = Article(label = "Chocolat", price = 12.0),
         statusCode = ArticleWrapperStatusType.DONE.code
     )
 
     val articleWrapper4 = ArticleWrapper(
-        article = Article(name = "Flan au coco", price = 7.0)
+        article = Article(label = "Flan au coco", price = 7.0)
     )
 
     val command1 = Command(
