@@ -1,9 +1,9 @@
 package com.jeanloth.project.android.kotlin.domain_models.entities
 
-data class Article(
-        val id : Long = 0,
-        val name : String,
-        val price : Double = 0.0,
-        var timeOrdered : Int = 0,
-        val category : Int = ArticleCategory.SALTED.code,
+import java.time.LocalDate
+
+data class Stock(
+    val id : Long = 0,
+    val lastUpdateTime : String = LocalDate.now().toString(),
+    var ingredientWrappers : List<IngredientWrapper> = mutableListOf(),
 )
