@@ -1,13 +1,13 @@
-package com.jeanloth.project.android.kotlin.domain.usescases.usecases.productWrapper
+package com.jeanloth.project.android.kotlin.domain.usescases.usecases.ingredientWrapper
 
 import com.jeanloth.project.android.kotlin.data.contracts.IngredientWrapperContract
-import com.jeanloth.project.android.kotlin.domain_models.entities.ProductWrapper
+import com.jeanloth.project.android.kotlin.domain_models.entities.IngredientWrapper
 
 class DeleteIngredientWrapperUseCase(
     private val ingredientWrapperContract: IngredientWrapperContract
 ){
 
-    fun invoke(productWrapper: ProductWrapper) {
-        ingredientWrapperContract.deleteProductWrapper(productWrapper)
+    fun invoke(ingredientWrapper: IngredientWrapper) {
+        ingredientWrapperContract.deleteIngredientWrapper(ingredientWrapper)
     }
 }
