@@ -1,4 +1,4 @@
-package com.jeanloth.project.android.kotlin.domain.usescases.usecases.productWrapper
+package com.jeanloth.project.android.kotlin.domain.usescases.usecases.ingredientWrapper
 
 import com.jeanloth.project.android.kotlin.data.contracts.IngredientWrapperContract
 
@@ -6,6 +6,6 @@ class ObserveAllIngredientWrappersUseCase(
     private val ingredientWrapperContract: IngredientWrapperContract
 ) {
 
-    fun invoke() = ingredientWrapperContract.observeAllProductWrappers()
+    fun invoke() = ingredientWrapperContract.observeAllIngredientWrappers()
 
 }
