@@ -1,9 +1,9 @@
 package com.jeanloth.project.android.kotlin.local.database
 
-import com.jeanloth.project.android.kotlin.local.entities.ProductWrapperEntity
+import com.jeanloth.project.android.kotlin.local.entities.IngredientWrapperEntity
 import io.objectbox.Box
 import io.objectbox.BoxStore
 
-class ProductWrapperDAO constructor(boxStore: BoxStore) : BaseDAO<ProductWrapperEntity>() {
-    public override val box: Box<ProductWrapperEntity> = boxStore.boxFor(ProductWrapperEntity::class.java)
+class IngredientWrapperDAO constructor(boxStore: BoxStore) : BaseDAO<IngredientWrapperEntity>() {
+    public override val box: Box<IngredientWrapperEntity> = boxStore.boxFor(IngredientWrapperEntity::class.java)
 }
