@@ -1,9 +1,9 @@
 package com.jeanloth.project.android.kotlin.local.database
 
-import com.jeanloth.project.android.kotlin.local.entities.IngredientWrapperEntity
+import com.jeanloth.project.android.kotlin.local.entities.RecipeWrapperEntity
 import io.objectbox.Box
 import io.objectbox.BoxStore
 
-class IngredientWrapperDAO constructor(boxStore: BoxStore) : BaseDAO<IngredientWrapperEntity>() {
-    public override val box: Box<IngredientWrapperEntity> = boxStore.boxFor(IngredientWrapperEntity::class.java)
+class RecipeWrapperDAO constructor(boxStore: BoxStore) : BaseDAO<RecipeWrapperEntity>() {
+    public override val box: Box<RecipeWrapperEntity> = boxStore.boxFor(RecipeWrapperEntity::class.java)
 }
