@@ -1,9 +1,9 @@
 package com.jeanloth.project.android.kotlin.local.database
 
-import com.jeanloth.project.android.kotlin.local.entities.ProductWrapperEntity
+import com.jeanloth.project.android.kotlin.local.entities.StockEntity
 import io.objectbox.Box
 import io.objectbox.BoxStore
 
-class ProductWrapperDAO constructor(boxStore: BoxStore) : BaseDAO<ProductWrapperEntity>() {
-    public override val box: Box<ProductWrapperEntity> = boxStore.boxFor(ProductWrapperEntity::class.java)
+class StockDAO constructor(boxStore: BoxStore) : BaseDAO<StockEntity>() {
+    public override val box: Box<StockEntity> = boxStore.boxFor(StockEntity::class.java)
 }
